Major problems solved
Moved the isLoggedIn boolean to the client inner class so that each client has their own. Rather than the pipe server having only one which was causing many of the problems.

the other problem was that the client then needed to be passed around to any methods that needed to check if the client was logged in. But that was a fairly simple thing.

All the thanks for solving these problems goes to my brother Adam Williamson as he was the one who found the cause of the problems and the solutions. The same for the major problems in the Login client program.